feat(panel): Journals joins the Agents hub as its third tab
